Rename .reader.sdmx{json,xml} to .reader.{json,xml}
